+++ packages/browseros-agent/apps/agent/components/sidebar/SettingsSidebar.tsx
@@ -9,7 +9,6 @@ import {
   RotateCcw,
   Search,
   Server,
-  ShieldAlert,
   ShieldCheck,
 } from 'lucide-react'
 import type { FC } from 'react'
@@ -87,12 +86,6 @@ const primarySettingsSections: NavSection[] = [
         feature: Feature.ALPHA_FEATURES_SUPPORT,
       },
       { name: 'BrowserOS as MCP', to: '/settings/mcp', icon: Server },
-      {
-        name: 'ACL Rules',
-        to: '/settings/acl',
-        icon: ShieldAlert,
-        feature: Feature.ALPHA_FEATURES_SUPPORT,
-      },
       {
         name: 'Usage & Billing',
         to: '/settings/usage',
