--- packages/browseros-agent/apps/agent/lib/acl/storage.ts
@@ -1,7 +1,0 @@
-import type { AclRule } from '@browseros/shared/types/acl'
-import { storage } from '#imports'
-
-export const aclRulesStorage = storage.defineItem<AclRule[]>(
-  'local:acl-rules',
-  { fallback: [] },
-)
