+++ packages/browseros-agent/apps/server/src/api/services/mcp/mcp-server.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import type { AclRule } from '@browseros/shared/types/acl'
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
 import { SetLevelRequestSchema } from '@modelcontextprotocol/sdk/types.js'
 import type { Browser } from '../../../browser/browser'
@@ -23,7 +22,6 @@ export interface McpServiceDeps {
   browser: Browser
   executionDir: string
   resourcesDir: string
-  aclRules?: AclRule[]
   klavisRef?: KlavisProxyRef
   observer?: ToolExecutionObserver
   // Per-request default windowId from the X-BrowserOS-Default-Window-Id
@@ -56,7 +54,6 @@ export function createMcpServer(deps: McpServiceDeps): McpServer {
       workingDir: deps.executionDir,
       resourcesDir: deps.resourcesDir,
     },
-    aclRules: deps.aclRules,
     observer: deps.observer,
     defaultWindowId: deps.defaultWindowId,
   })
