+++ packages/browseros-agent/apps/server/src/api/types.ts
@@ -46,18 +46,6 @@ export const ChatRequestSchema = AgentLLMConfigSchema.extend({
   mode: z.enum(['chat', 'agent']).optional().default('agent'),
   origin: z.enum(['sidepanel', 'newtab']).optional().default('sidepanel'),
   declinedApps: z.array(z.string()).optional(),
-  aclRules: z
-    .array(
-      z.object({
-        id: z.string(),
-        sitePattern: z.string(),
-        selector: z.string().optional(),
-        textMatch: z.string().optional(),
-        description: z.string().optional(),
-        enabled: z.boolean(),
-      }),
-    )
-    .optional(),
   toolApprovalConfig: z
     .object({
       categories: z.record(z.boolean()),
