+++ packages/browseros-agent/apps/server/src/browser/browser.ts
@@ -1,5 +1,4 @@
 import type { ProtocolApi } from '@browseros/cdp-protocol/protocol-api'
-import type { ElementProperties } from '@browseros/shared/types/acl'
 import { logger } from '../lib/logger'
 import type { CdpBackend } from './backends/types'
 import type { BookmarkNode } from './bookmarks'
@@ -91,24 +90,6 @@ const EXCLUDED_URL_PREFIXES = [
   'chrome-search://',
   'devtools://',
 ]
-
-const ACTIONABLE_SELECTOR = [
-  'button',
-  'a[href]',
-  'input',
-  'select',
-  'textarea',
-  'summary',
-  '[role="button"]',
-  '[role="link"]',
-  '[role="checkbox"]',
-  '[role="radio"]',
-  '[role="switch"]',
-  '[role="tab"]',
-  '[role="option"]',
-  '[onclick]',
-  '[tabindex]',
-].join(',')
 
 export class Browser {
   private cdp: CdpBackend
@@ -295,207 +276,6 @@ export class Browser {
     const sessionId = this.sessions.get(info.targetId)
     if (!sessionId) return null
     return this.cdp.session(sessionId)
-  }
-
-  async resolveActionableElement(
-    pageId: number,
-    backendNodeId: number,
-  ): Promise<number | null> {
-    const session = await this.resolveSession(pageId)
-    try {
-      const resolved = await session.DOM.resolveNode({ backendNodeId })
-      const objectId = resolved.object?.objectId
-      if (!objectId) return backendNodeId
-
-      const actionable = await session.Runtime.callFunctionOn({
-        functionDeclaration: `function(selector){
-          var element = this instanceof Element
-            ? this
-            : this && this.parentElement
-              ? this.parentElement
-              : this && this.parentNode instanceof Element
-                ? this.parentNode
-                : null;
-          if (!element) return null;
-          return element.closest(selector) || element;
-        }`,
-        objectId,
-        arguments: [{ value: ACTIONABLE_SELECTOR }],
-      })
-
-      const actionableObjectId = actionable.result?.objectId
-      if (!actionableObjectId) return backendNodeId
-
-      const desc = await session.DOM.describeNode({
-        objectId: actionableObjectId,
-      })
-      return desc.node?.backendNodeId ?? backendNodeId
-    } catch {
-      return null
-    }
-  }
-
-  async resolveElementAtPoint(
-    pageId: number,
-    x: number,
-    y: number,
-  ): Promise<number | null> {
-    const session = await this.resolveSession(pageId)
-    try {
-      const fromDom = await session.Runtime.evaluate({
-        expression: `document.elementFromPoint(${Math.round(x)}, ${Math.round(y)})`,
-      })
-      const objectId = fromDom.result?.objectId
-      if (objectId) {
-        const desc = await session.DOM.describeNode({ objectId })
-        const backendNodeId = desc.node?.backendNodeId
-        if (backendNodeId) {
-          return await this.resolveActionableElement(pageId, backendNodeId)
-        }
-      }
-    } catch {
-      // fall through to CDP hit-testing
-    }
-
-    try {
-      const located = await session.DOM.getNodeForLocation({
-        x: Math.round(x),
-        y: Math.round(y),
-        includeUserAgentShadowDOM: true,
-        ignorePointerEventsNone: true,
-      })
-      return await this.resolveActionableElement(pageId, located.backendNodeId)
-    } catch {
-      return null
-    }
-  }
-
-  async resolveElementProperties(
-    pageId: number,
-    backendNodeId: number,
-  ): Promise<ElementProperties | null> {
-    const session = await this.resolveSession(pageId)
-    try {
-      const targetNodeId =
-        (await this.resolveActionableElement(pageId, backendNodeId)) ??
-        backendNodeId
-      const desc = await session.DOM.describeNode({
-        backendNodeId: targetNodeId,
-        depth: 0,
-      })
-      const node = desc.node
-      const attrs = parseNodeAttributes(node)
-
-      const resolved = await session.DOM.resolveNode({
-        backendNodeId: targetNodeId,
-      })
-      const objectId = resolved.object?.objectId
-      let textContent = ''
-      let labelText = ''
-      if (objectId) {
-        const textResult = await session.Runtime.callFunctionOn({
-          functionDeclaration: `function(){
-            var text = (this.innerText || this.textContent || '').trim();
-            var aria = this.getAttribute('aria-label') || '';
-            var placeholder = this.getAttribute('placeholder') || '';
-            var title = this.getAttribute('title') || '';
-            var value = typeof this.value === 'string' ? this.value : '';
-            var labels = Array.from(this.labels || [])
-              .map(function(label){ return (label.innerText || label.textContent || '').trim(); })
-              .filter(Boolean)
-              .join(' ');
-            return {
-              textContent: text.substring(0, 200),
-              labelText: [aria, labels, placeholder, title, value, text]
-                .filter(Boolean)
-                .join(' ')
-                .trim()
-                .substring(0, 400),
-            };
-          }`,
-          objectId,
-          returnByValue: true,
-        })
-        const value = (textResult.result?.value ?? {}) as {
-          textContent?: string
-          labelText?: string
-        }
-        textContent = value.textContent ?? ''
-        labelText = value.labelText ?? ''
-      }
-
-      return {
-        tagName: node.localName ?? '',
-        textContent,
-        attributes: attrs,
-        labelText,
-        ariaLabel: attrs['aria-label'],
-        role: attrs.role,
-      }
-    } catch {
-      return null
-    }
-  }
-
-  async highlightBlockedElement(
-    pageId: number,
-    backendNodeId: number,
-    reason: string,
-  ): Promise<void> {
-    const session = await this.resolveSession(pageId)
-    const targetNodeId =
-      (await this.resolveActionableElement(pageId, backendNodeId)) ??
-      backendNodeId
-
-    try {
-      const resolved = await session.DOM.resolveNode({
-        backendNodeId: targetNodeId,
-      })
-      const objectId = resolved.object?.objectId
-      if (!objectId) return
-
-      await session.Runtime.callFunctionOn({
-        functionDeclaration: `function(reason){
-          var existing = document.getElementById('__browseros_acl_block_overlay');
-          if (existing) existing.remove();
-          var existingStyle = document.getElementById('__browseros_acl_block_style');
-          if (!existingStyle) {
-            var style = document.createElement('style');
-            style.id = '__browseros_acl_block_style';
-            style.textContent = [
-              '#__browseros_acl_block_overlay{position:absolute;pointer-events:none;z-index:2147483647;}',
-              '#__browseros_acl_block_overlay .ring{position:absolute;inset:0;border:2px solid rgba(220,38,38,0.95);background:rgba(220,38,38,0.14);border-radius:10px;box-shadow:0 0 0 3px rgba(255,255,255,0.75);}',
-              '#__browseros_acl_block_overlay .badge{position:absolute;top:-10px;right:-10px;background:rgba(153,27,27,0.96);color:white;font:600 11px/1.2 system-ui,sans-serif;padding:6px 8px;border-radius:999px;white-space:nowrap;box-shadow:0 6px 18px rgba(0,0,0,0.2);}',
-            ].join('');
-            document.head.appendChild(style);
-          }
-          var rect = this.getBoundingClientRect();
-          if (!rect.width || !rect.height) return;
-          var overlay = document.createElement('div');
-          overlay.id = '__browseros_acl_block_overlay';
-          overlay.style.left = (rect.left + window.scrollX) + 'px';
-          overlay.style.top = (rect.top + window.scrollY) + 'px';
-          overlay.style.width = rect.width + 'px';
-          overlay.style.height = rect.height + 'px';
-          var ring = document.createElement('div');
-          ring.className = 'ring';
-          var badge = document.createElement('div');
-          badge.className = 'badge';
-          badge.textContent = reason || 'Blocked';
-          overlay.appendChild(ring);
-          overlay.appendChild(badge);
-          document.body.appendChild(overlay);
-          window.setTimeout(function(){
-            var current = document.getElementById('__browseros_acl_block_overlay');
-            if (current) current.remove();
-          }, 2500);
-        }`,
-        objectId,
-        arguments: [{ value: reason }],
-      })
-    } catch {
-      // best-effort visual feedback
-    }
   }
 
   async resolveTabIds(tabIds: number[]): Promise<Map<number, number>> {
