+++ packages/browseros-agent/apps/server/src/tools/acl/README.md
@@ -1,10 +1,10 @@
-# ACL Matcher
+# ACL Scorer Reference
 
-The ACL matcher blocks guarded tool actions (click, fill, hover, etc.) when they target elements that match user-defined access control rules. It scores each rule against the target element using a combination of exact, fuzzy, and semantic similarity — then blocks if the confidence exceeds a threshold.
+This folder now contains reference-only ACL scoring code. BrowserOS no longer exposes ACL rule UI, APIs, shared contracts, or production tool enforcement. Nothing in the server runtime should import or call the scorer.
 
 ## How it works
 
-When a guarded tool is invoked, `acl-guard.ts` resolves the target element's properties (text content, aria labels, attributes, etc.) and runs them through the scoring pipeline:
+`acl-scorer.ts` can still score a captured element fixture against rule-like inputs using the old matching pipeline:
 
 1. **Site filtering** — rules are filtered to those matching the current page URL
 2. **Site-only rules** — rules with no selector/text/description block the entire site immediately
@@ -16,21 +16,16 @@ When a guarded tool is invoked, `acl-guard.ts` resolves the target element's pro
 | Fuzzy | 25% | Edit distance ratio between rule terms and element text windows |
 | Semantic | 50% | Cosine similarity of sentence embeddings (BAAI/bge-small-en-v1.5 via ONNX) |
 
-The weighted scores produce a **confidence** value between 0 and 1. If confidence >= **0.4** (Handpicked, needs updating), the action is blocked.
+The weighted scores produce a **confidence** value between 0 and 1. If confidence >= **0.4**, the scorer marks the fixture decision as blocked. This is only a reference decision, not runtime enforcement.
 
 ## Files
 
 | File | Purpose |
 |------|---------|
-| `acl-guard.ts` | Entry point — called by `framework.ts` during tool execution |
-| `acl-scorer.ts` | Core pipeline: text normalization, feature extraction, scoring, decision |
+| `acl-scorer.ts` | Reference pipeline: local ACL-like types, site pattern matching, feature extraction, scoring, decision |
 | `acl-embeddings.ts` | Lazy-loaded `@huggingface/transformers` pipeline for semantic similarity |
 | `acl-edit-distance.ts` | Levenshtein edit distance ratio for fuzzy matching |
 | `acl-stopwords.ts` | Static set of 198 English stopwords (from NLTK corpus) |
-
-Shared types and basic matchers live in `packages/shared/`:
-- `src/types/acl.ts` — `AclRule` and `ElementProperties` interfaces
-- `src/acl/match.ts` — site pattern globbing and CSS selector matching
 
 ## Embedding model
 
@@ -41,15 +36,5 @@ Override the model with the `ACL_EMBEDDING_MODEL` environment variable (e.g. `AC
 ## Testing
 
 ```bash
-bun --env-file=.env.development test apps/server/tests/tools/acl-scorer.test.ts
+bun --env-file=apps/server/.env.development test apps/server/tests/tools/acl-scorer.test.ts
 ```
-
-Test fixtures live in `apps/server/tests/__fixtures__/acl/` (courtesy of claude code):
-
-| Fixture | Tests |
-|---------|-------|
-| `submit-button.json` | Exact match — "Place Order" button vs "block checkout submit" rule |
-| `semantic-payment.json` | Semantic match — "Proceed to Checkout" vs "prevent purchase actions" |
-| `semantic-delete.json` | Semantic match — "Remove my account permanently" vs "block account deletion" |
-| `semantic-send-email.json` | Semantic match — send button vs "do not dispatch emails" |
-| `semantic-safe.json` | False positive — "View Report" should NOT be blocked by payment/delete rules |
