+++ packages/browseros-agent/apps/server/src/tools/framework.ts
@@ -1,7 +1,6 @@
 import { tmpdir } from 'node:os'
 import { resolve } from 'node:path'
 import type { ToolApprovalCategoryId } from '@browseros/shared/constants/tool-approval'
-import type { AclRule } from '@browseros/shared/types/acl'
 import type { z } from 'zod'
 import type { Browser } from '../browser/browser'
 import { ToolResponse, type ToolResult } from './response'
@@ -35,7 +34,6 @@ export type ToolContext = {
   browser: Browser
   directories: ToolDirectories
   session?: ToolSessionContext
-  aclRules?: AclRule[]
 }
 
 export function resolveWorkingPath(
@@ -98,34 +96,6 @@ export async function executeTool(
   if (signal.aborted) {
     response.error('Request was aborted')
     return response.toResult()
-  }
-
-  if (ctx.aclRules?.length) {
-    const { checkAcl } = await import('./acl/acl-guard')
-    const check = await checkAcl(
-      tool.name,
-      args as Record<string, unknown>,
-      ctx.browser,
-      ctx.aclRules,
-    )
-    if (check.blocked) {
-      const desc =
-        check.rule?.description ??
-        check.rule?.textMatch ??
-        check.rule?.sitePattern ??
-        'ACL rule'
-      if (check.pageId !== undefined && check.elementId !== undefined) {
-        await ctx.browser.highlightBlockedElement(
-          check.pageId,
-          check.elementId,
-          desc,
-        )
-      }
-      response.error(
-        `Action blocked by ACL rule: "${desc}". The element on this page is restricted. Choose a different action or skip this step.`,
-      )
-      return response.toResult()
-    }
   }
 
   try {
