+++ packages/browseros-agent/packages/shared/package.json
@@ -76,14 +76,6 @@
     "./sentry/sanitize": {
       "types": "./src/sentry/sanitize.ts",
       "default": "./src/sentry/sanitize.ts"
-    },
-    "./types/acl": {
-      "types": "./src/types/acl.ts",
-      "default": "./src/types/acl.ts"
-    },
-    "./acl/match": {
-      "types": "./src/acl/match.ts",
-      "default": "./src/acl/match.ts"
     }
   }
 }
