--- packages/browseros-agent/packages/shared/src/types/acl.ts
@@ -1,17 +1,0 @@
-export interface AclRule {
-  id: string
-  sitePattern: string
-  selector?: string
-  textMatch?: string
-  description?: string
-  enabled: boolean
-}
-
-export interface ElementProperties {
-  tagName: string
-  textContent: string
-  attributes: Record<string, string>
-  labelText?: string
-  ariaLabel?: string
-  role?: string
-}
